\name{CANOVA}
\alias{CANOVA}

\title{Continuous Analysis of Variance test for dependence}

\description{
CANOVA test for dependence. Note that the size of neighborhood K = 2 following the default in the cited publication.
}

\references{
Wang, Y., Li, Y., Cao, H., Xiong, M., Shugart, Y. Y., & Jin, L. (2015). Efficient test for nonlinear dependence of two continuous variables. BMC bioinformatics, 16(1), 1.
}

