\name{EL}
\alias{EL}

\title{Empirical Likelihood based test for dependence}

\description{
Empirical Likelihood based test for dependence. See references.
}

\references{
Einmahl, J. H., & McKeague, I. W. (2003). Empirical likelihood based hypothesis testing. Bernoulli, 267-290.
}



