\name{Hoeffding}
\alias{Hoeffding}

\title{Hoeffding's test for dependence}
\description{
   Test statistic is computed by hoeffd\{Hmisc\}. See \link[Hmisc]{hoeffd}.
   Note that test statistic D is 30 times the original test statistic in the original publication.
}

\references{
Harrell Jr FE, Dupont MC (2006). "The Hmisc Package." R package version, 3, 0-12.
}
