\name{LSAT}
\alias{LSAT}
\docType{data}
\title{
LSAT dataset
}
\description{
A dataset of average law school admission test (LSAT) and grade point average
(GPA) from 82 American law schools participated in a large study of admission practices.
}
\usage{data("LSAT")}
\format{
  A data frame with 82 observations on the following 3 variables.
  \describe{
    \item{\code{School}}{a numeric vector of school numbers.}
    \item{\code{LSAT}}{a numeric vector of LSAT's.}
    \item{\code{GPA}}{a numeric vector of GPA's.}
  }
}
\details{
details see references.
}
\source{
Efron B, Tibshirani RJ (1994). An Introduction to the Bootstrap. CRC Press.
}
\references{
Efron B, Tibshirani RJ (1994). An Introduction to the Bootstrap. CRC Press.
}

\keyword{datasets}
