\name{Vexler}
\alias{Vexler}

\title{Vexler's test for dependence}

\description{
A method based on empirical likelihood ratio test. Published by Dr. Vexler in 2014. See reference.

}

\references{
Vexler A, Tsai WM, Hutson AD (2014). A Simple Density-Based Empirical Likelihood Ratio
Test for Independence."
}
